Added test for SpringBasedLoaderAction, with test command on jar.
