+++ interaction-dynamic-loader/src/test/java/com/temenos/interaction/loader/detector/AnnotationBasedCommandControllerTest.java
@@ -1,5 +1,26 @@
 package com.temenos.interaction.loader.detector;
 
+/*
+ * #%L
+ * interaction-dynamic-loader
+ * %%
+ * Copyright (C) 2012 - 2015 Temenos Holdings N.V.
+ * %%
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ * 
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ * 
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * #L%
+ */
+
 import com.temenos.interaction.core.command.AnnotationBasedCommandController;
 import com.temenos.interaction.core.command.InteractionCommand;
 import com.temenos.interaction.loader.classloader.ParentLastURLClassloader;
