+++ interaction-dynamic-loader/src/test/java/com/temenos/interaction/loader/detector/stub/DirectoryEventInterestedActionTestStub.java
@@ -1,0 +1,54 @@
+package com.temenos.interaction.loader.detector.stub;
+
+/*
+ * #%L
+ * interaction-dynamic-loader
+ * %%
+ * Copyright (C) 2012 - 2015 Temenos Holdings N.V.
+ * %%
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ * 
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ * 
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * #L%
+ */
+
+import com.temenos.interaction.core.loader.Action;
+import com.temenos.interaction.core.loader.FileEvent;
+import java.io.File;
+
+/**
+ *
+ * @author andres
+ */
+public class DirectoryEventInterestedActionTestStub implements Action<FileEvent<File>> {
+    
+    private int callCount = 0;
+    private FileEvent<File> lastEvent;
+    
+    @Override
+    public void execute(FileEvent<File> event) {
+       callCount++;
+       lastEvent = event;
+    }
+    
+    public int getCallCount() {
+        return callCount;
+    }
+
+    /**
+     * @return the lastEvent
+     */
+    public FileEvent<File> getLastEvent() {
+        return lastEvent;
+    }
+    
+}
