+++ interaction-dynamic-loader/src/test/resources/spring-test-contexts/directory-loader-test-context.xml
@@ -28,12 +28,13 @@
 		http://www.springframework.org/schema/context
 		http://www.springframework.org/schema/context/spring-context-3.0.xsd">
     
-    <bean id="test-action" class="com.temenos.interaction.loader.detector.DirectoryEventInterestedActionTestStub"/>
+    <bean id="test-action" class="com.temenos.interaction.loader.detector.stub.DirectoryEventInterestedActionTestStub"/>
     <bean id="directory-watcher" class="com.temenos.interaction.loader.detector.DirectoryChangeActionNotifier">
         <property name="listeners">
             <list>
                 <ref bean="test-action"/>
             </list>
         </property>
+        <property name="intervalSeconds" value="2"/>
     </bean>
 </beans>
